+++ exporters/otlp/common/src/main/java/io/opentelemetry/exporter/internal/otlp/BoolAnyValueStatelessMarshaler.java
@@ -12,6 +12,7 @@ import io.opentelemetry.exporter.internal.marshal.StatelessMarshaler;
 import io.opentelemetry.proto.common.v1.internal.AnyValue;
 import java.io.IOException;
 
+/** See {@link BoolAnyValueMarshaler}. */
 final class BoolAnyValueStatelessMarshaler implements StatelessMarshaler<Boolean> {
   static final BoolAnyValueStatelessMarshaler INSTANCE = new BoolAnyValueStatelessMarshaler();
