+++ exporters/otlp/common/src/main/java/io/opentelemetry/exporter/internal/otlp/IntAnyValueStatelessMarshaler.java
@@ -12,6 +12,7 @@ import io.opentelemetry.exporter.internal.marshal.StatelessMarshaler;
 import io.opentelemetry.proto.common.v1.internal.AnyValue;
 import java.io.IOException;
 
+/** See {@link IntAnyValueMarshaler}. */
 final class IntAnyValueStatelessMarshaler implements StatelessMarshaler<Long> {
   static final IntAnyValueStatelessMarshaler INSTANCE = new IntAnyValueStatelessMarshaler();
