+++ exporters/otlp/common/src/main/java/io/opentelemetry/exporter/internal/otlp/StringAnyValueStatelessMarshaler.java
@@ -6,14 +6,14 @@
 package io.opentelemetry.exporter.internal.otlp;
 
 import io.opentelemetry.exporter.internal.marshal.MarshalerContext;
-import io.opentelemetry.exporter.internal.marshal.MarshalerUtil;
 import io.opentelemetry.exporter.internal.marshal.Serializer;
 import io.opentelemetry.exporter.internal.marshal.StatelessMarshaler;
+import io.opentelemetry.exporter.internal.marshal.StatelessMarshalerUtil;
 import io.opentelemetry.proto.common.v1.internal.AnyValue;
 import java.io.IOException;
 
 /**
- * A Marshaler of string-valued {@link AnyValue}.
+ * A Marshaler of string-valued {@link AnyValue}. See {@link StringAnyValueMarshaler}.
  *
  * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
  * at any time.
@@ -29,6 +29,6 @@ final class StringAnyValueStatelessMarshaler implements StatelessMarshaler<Strin
 
   @Override
   public int getBinarySerializedSize(String value, MarshalerContext context) {
-    return MarshalerUtil.sizeString(AnyValue.STRING_VALUE, value, context);
+    return StatelessMarshalerUtil.sizeString(AnyValue.STRING_VALUE, value, context);
   }
 }
