+++ exporters/otlp/common/src/main/java/io/opentelemetry/exporter/internal/otlp/traces/ResourceSpansStatelessMarshaler.java
@@ -9,6 +9,7 @@ import io.opentelemetry.exporter.internal.marshal.MarshalerContext;
 import io.opentelemetry.exporter.internal.marshal.MarshalerUtil;
 import io.opentelemetry.exporter.internal.marshal.Serializer;
 import io.opentelemetry.exporter.internal.marshal.StatelessMarshaler2;
+import io.opentelemetry.exporter.internal.marshal.StatelessMarshalerUtil;
 import io.opentelemetry.exporter.internal.otlp.ResourceMarshaler;
 import io.opentelemetry.proto.trace.v1.internal.ResourceSpans;
 import io.opentelemetry.sdk.common.InstrumentationScopeInfo;
@@ -19,7 +20,7 @@ import java.util.List;
 import java.util.Map;
 
 /**
- * A Marshaler of ResourceSpans.
+ * A Marshaler of ResourceSpans. See {@link ResourceSpansMarshaler}.
  *
  * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
  * at any time.
@@ -63,14 +64,16 @@ public final class ResourceSpansStatelessMarshaler
     size += MarshalerUtil.sizeMessage(ResourceSpans.RESOURCE, resourceMarshaler);
 
     size +=
-        MarshalerUtil.sizeRepeatedMessage(
+        StatelessMarshalerUtil.sizeRepeatedMessage(
             ResourceSpans.SCOPE_SPANS,
             scopeMap,
             InstrumentationScopeSpansStatelessMarshaler.INSTANCE,
             context,
             SCOPE_SPAN_SIZE_CALCULATOR_KEY);
 
-    size += MarshalerUtil.sizeString(ResourceSpans.SCHEMA_URL, resource.getSchemaUrl(), context);
+    size +=
+        StatelessMarshalerUtil.sizeString(
+            ResourceSpans.SCHEMA_URL, resource.getSchemaUrl(), context);
 
     return size;
   }
